Add files via upload
I looked at Isaac's R analysis for his branch, and the information in it seems like practically everything we need for tomorrow, so I just did some editing to the formatting and some spell-checking (I know it wasn't necessarily important since milage is an alternate spelling of mileage, and that can be changed back if that's better). I also added the summaries for the data sets before and after the cleaning that's also discussed in the analysis to add context.